Clear previously set values not present in the current file snapshot
Value is set in previous file, but it is not present in current file
snapshot. This might happen if device was switched off and the value was
cleared during device inactivity. For example TC got free whilst cabinet
was without power.
We have to append this lost information after the current file snapshot,
to reflect cleared value.