Move RpcValue metatype header include
Using RpcCall needs rpc.h inclusion (even in the code that uses it.)
Almost everyone who uses RpcCall will have socketrpcconnection.h
included (possibly through a chain of includes), so I'll move the
include to the header file.

There is a possibility that the user has a ClientConnection pointer and
doesn't have the definition included. In that situation, the user could
use RpcCall and NOT have rpc.h included. I have decided that this
situation will never happen (the user wouldn't be able to do much else
with the ClientConnection pointer).