Fix RpcValue metatype header include
Everyone who uses a signal that has an RpcValue parameter might use it
for a QueuedConnection. For that, the user needs to have "rpc.h"
included somewhere, and we need to supply it to them in *some* header.

The user must see the declaration of the signal, therefore, the user
must include the appropriate header. There are only two headers with
RpcValue signals, socketrpcconnection.h and rpcresponsecallback.h. I've
already edited the first one, so this patch fixes the second one.

This way, everyone who uses an RpcValue signal, will also see "rpc.h".