Add support for qt6 (#200)
* qt6: cmake: Enable C++17

* qt6: cmake: Add support for Qt6

* qt6: Fix forward declaration

QStringList in Qt6 is now just a typedef to QList<QString> and it is now
forward-declared by Qt at this location.

* qt6: Include rpc.h for metatype declarations

In Qt6, the QMetaTypeId struct is implicitly instantiated in the MOC for
this file. I'm not exactly sure why. Fix this by including the metatype
delcaration header.

* qt6: Get rid of some QSerialPort errors

The documentation says that these are obsolete.

* qt6: Add compat lib

QStringRef is deprecated, but I'm going the least amount of work right
now.

* qt6: Change QMatrix to QTransform

* qt6: Add explicit cast to assignment of QChar

I suspect that Qt5 had an explicit assignment operator for =.

* qt6: Fix string concatenation

* qt6: Add a printer for QStringView

It seems that QStringView could convert to string ref in Qt5, but in Qt6
it can't (for logical reasons).

* qt6: Get rid of QDesktopWidget

I'm not sure if the fix is correct, but the doc for grabWindow says I
should get the winId from a QWidget. I'm not really sure which one would
that be, so I just used `this`.

* qt6: Get rid of QVariant::Type

About CLIOptions::value_helper: it seems that creating the value in this
way is "flawed" and shouldn't probably ever be done. There is not
replacement in Qt6.